Refactor auth module on the backend
