+++ backend/src/core/auth/auth.service.ts
@@ -1,24 +1,10 @@
-import type { FastifyInstance, FastifyReply, FastifyRequest } from "fastify";
+import type { FastifyInstance } from "fastify";
 import { Result, err, ok } from "neverthrow";
 import bcrypt from "bcrypt";
 import QRCode from "qrcode";
 import * as speakeasy from "speakeasy";
 import { type JwtPayload, TotpSetupPayload, userSchemas } from "@darrenkuro/pong-core";
-import { ApiError } from "../../utils/api-response.ts";
 import { UserRecord } from "../db/db.types.ts";
-
-export const verifyCookie = async (req: FastifyRequest, _: FastifyReply) => {
-    const { cookieName } = req.server.config;
-    const token = req.cookies?.[cookieName];
-    if (!token) return;
-
-    const payload = req.server.authService.verifyJwtToken(token);
-    if (payload.isOk() && payload.value.id && !isNaN(Number(payload.value.id))) {
-        req.userId = Number(payload.value.id);
-        req.username = payload.value.username;
-        req.userDisplayName = payload.value.displayName;
-    }
-};
 
 export const createAuthService = (app: FastifyInstance) => {
     const { jwt } = app;
@@ -42,14 +28,14 @@ export const createAuthService = (app: FastifyInstance) => {
         return jwt.sign(payload, { expiresIn: exp });
     };
 
-    const verifyJwtToken = (token: string): Result<JwtPayload, Error> => {
+    const verifyJwtToken = (token: string): Result<JwtPayload, string> => {
         try {
             const payload = jwt.verify(token) as JwtPayload;
             // Runtime type check to ensure token is valid and has the correct fields
             userSchemas.jwtPayload.parse(payload);
             return ok(payload);
         } catch (error) {
-            return err(new Error("Invalid JWT token or payload"));
+            return err("Invalid JWT token or payload");
         }
     };
 
@@ -67,14 +53,6 @@ export const createAuthService = (app: FastifyInstance) => {
         return speakeasy.totp.verify({ secret, token, encoding });
     };
 
-    const requireAuth = async (req: FastifyRequest, reply: FastifyReply) => {
-        // Check that userId exists and is larger than 0 (attached by onRequest hook)
-        if (!req.userId || req.userId <= 0) {
-            const error = new ApiError("UNAUTHORIZED", 401, "Authentication required");
-            return error.send(reply);
-        }
-    };
-
     return {
         hashPassword,
         comparePassword,
@@ -82,6 +60,5 @@ export const createAuthService = (app: FastifyInstance) => {
         verifyJwtToken,
         generateTotpSecret,
         verifyTotpToken,
-        requireAuth,
     };
 };
